requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
h11==0.13.0
idna==3.3
mss==6.1.0
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.28.0
selenium==4.2.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.4
websockets==10.3
wincertstore==0.2
wsproto==1.1.0
